+++ ui/src/kapacitor/constants/index.ts
@@ -34,6 +34,7 @@ export enum AlertTypes {
   servicenow = 'servicenow',
   bigpanda = 'bigpanda',
   teams = 'teams',
+  zenoss = 'zenoss',
 }
 
 export enum AlertDisplayText {
@@ -53,6 +54,7 @@ export enum AlertDisplayText {
   servicenow = 'ServiceNow',
   bigpanda = 'BigPanda',
   teams = 'Teams',
+  zenoss = 'Zenoss',
 }
 
 export const SupportedServices: string[] = [
@@ -72,6 +74,7 @@ export const SupportedServices: string[] = [
   'teams',
   'telegram',
   'victorops',
+  'zenoss',
 ]
 
 export const defaultRuleConfigs = {
@@ -219,6 +222,7 @@ export const MAP_KEYS_FROM_CONFIG: KeyMappings = {
   servicenow: 'serviceNow',
   bigpanda: 'bigPanda',
   teams: 'teams',
+  zenoss: 'zenoss',
 }
 
 // ALERTS_FROM_CONFIG the array of fields to accept from Kapacitor Config
@@ -250,6 +254,7 @@ export const ALERTS_FROM_CONFIG: FieldsFromConfigAlerts = {
   teams: [
     /* 'channel-url' is intentionally skipped */
   ],
+  zenoss: ['action', 'method', '_type', 'tid'],
 }
 
 export const MAP_FIELD_KEYS_FROM_CONFIG: ConfigKeyMaps = {
@@ -276,6 +281,7 @@ export const MAP_FIELD_KEYS_FROM_CONFIG: ConfigKeyMaps = {
   serviceNow: {},
   bigPanda: {},
   teams: {},
+  zenoss: {},
 }
 
 // HANDLERS_TO_RULE returns array of fields that may be updated for each alert on rule.
@@ -321,4 +327,14 @@ export const HANDLERS_TO_RULE_THEM_ALL: FieldsFromAllAlerts = {
   ],
   bigPanda: ['app-key', 'primary-property', 'secondary-property'],
   teams: ['channel-url'],
+  zenoss: [
+    'action',
+    'method',
+    '_type',
+    'tid',
+    'device',
+    'component',
+    'evclasskey',
+    'evclass',
+  ],
 }